refactor(knowledge): Move useKnowledgeItemMetadata to utils/knowledge.ts module
